Add FlowObject Strategy.
