+++ src/domain-services/flows/strategy/flow-search-strategy.ts
@@ -1,5 +1,12 @@
 import { Database } from '@unocha/hpc-api-core/src/db';
 import { FlowSearchResult } from '../graphql/types';
+import { InstanceOfModel } from '@unocha/hpc-api-core/src/db/util/types';
+import { InstanceDataOfModel } from '@unocha/hpc-api-core/src/db/util/raw-model';
+
+export interface FlowSearchStrategyResponse {
+  flows: InstanceDataOfModel<Database['flow']>[];
+  count: number;
+}
 
 export interface FlowSearchStrategy {
   search(
@@ -8,5 +15,5 @@ export interface FlowSearchStrategy {
     limit: number,
     cursorCondition: any,
     models: Database
-  ): Promise<FlowSearchResult>;
+  ): Promise<FlowSearchStrategyResponse>;
 }
